refactor: 작성한 코드를 역할 단위로 함수 및 파일을 분리
