+++ README.md
@@ -28,7 +28,7 @@
 
 ### 5. 추가 작업
 
-- [ ] 작성한 코드를 역할 단위로 분리한다.
+- [-] 작성한 코드를 역할 단위로 분리한다.
 - [ ] 테스트 코드를 추가하여 다양한 입력을 확인하고 보완한다.
 
 ### 입출력 요구 사항
